startkde sets QQC2 style to org.kde.desktop, pull it in
(Bhushan is not sure if this is needed on mobile, but seeing as there is
no mobile style it is probably needed for now as otherwise kirigami2 is
properly breeze styled. At any rate there aren't many options here.
Should this turn out entirely unwanted the dependency can be moved to
plasma-desktop instead, where we already pull in stack-dependencies to
ensure a consistent desktop, so while this dep doesn't belong there it
will be fine all the same)